Rename external services and add price getter
