+++ src/components/chats/ChatItem/RepliedMessagePreview.tsx
@@ -2,7 +2,7 @@ import ImageLoader from '@/components/ImageLoader'
 import Name from '@/components/Name'
 import useRandomColor from '@/hooks/useRandomColor'
 import { getPostQuery } from '@/services/api/query'
-import { getNftDataQuery } from '@/services/moralis/query'
+import { getNftDataQuery } from '@/services/external/query'
 import { cx } from '@/utils/class-names'
 import { truncateText } from '@/utils/strings'
 import { ComponentProps, useState } from 'react'
