+++ src/components/chats/ChatRoom/RepliedMessage.tsx
@@ -3,7 +3,7 @@ import ImageLoader from '@/components/ImageLoader'
 import Name from '@/components/Name'
 import useRandomColor from '@/hooks/useRandomColor'
 import { getPostQuery } from '@/services/api/query'
-import { getNftDataQuery } from '@/services/moralis/query'
+import { getNftDataQuery } from '@/services/external/query'
 import { useMessageData } from '@/stores/message'
 import { cx } from '@/utils/class-names'
 import { ComponentProps } from 'react'
