+++ src/components/extensions/nft/NftAttachmentModal.tsx
@@ -5,7 +5,7 @@ import Input, { InputProps } from '@/components/inputs/Input'
 import { ModalFunctionalityProps } from '@/components/modals/Modal'
 import useAutofocus from '@/hooks/useAutofocus'
 import useDebounce from '@/hooks/useDebounce'
-import { getNftDataQuery } from '@/services/moralis/query'
+import { getNftDataQuery } from '@/services/external/query'
 import { NftProperties } from '@subsocial/api/types'
 import { useEffect, useState } from 'react'
 import { HiTrash } from 'react-icons/hi2'
